feat: Home-page-layout & refactoring of project
