+++ src/Components/Common/Cards.scss
@@ -25,17 +25,20 @@ body {
     &--awap {
         flex: 1 1 auto;
     }
+    &--right-float {
+        float: right;
+    }
     &--uppercase {
         text-transform: uppercase;
     }
     &--secondary {
         background: #000;
-        color: #fff;
+        color: rgb(255, 255, 255);
         box-shadow: 0 2px 4px 0 rgba(0, 28, 61, 0.3);
     }
     &--tertiary {
         background: #1e79e4;
-        color: #fff;
+        color: rgb(17, 16, 16);
         box-shadow: 0 2px 4px 0 rgba(30, 121, 228, 0.3);
     }
 }
@@ -61,6 +64,10 @@ body {
     &__heading {
         color: #001c3d;
     }
+    &__description {
+        color: #001c3d;
+        text-align: right;
+    }
     &__bottom {
         background: #fff;
         border: solid 1px #000;
@@ -73,7 +80,7 @@ body {
         border: solid 1px #000;
         display: flex;
         padding: 10px 10px 10px;
-        justify-content: center;
+        // justify-content: center;
         // padding-left: 5px;
         //padding-right: 5px;
         // margin-left: -10px;
