+++ src/Components/Common/Footer.scss
@@ -4,6 +4,7 @@
 
 body {
     margin: 0;
+    bottom: 0px;
 }
 
 footer {
@@ -18,6 +19,7 @@ footer {
     font-size: 1.2rem;
     float: center;
 }
+
 .footer__box {
     display: block;
     //padding: 1em;
