+++ src/Components/Common/Navbar.scss
@@ -52,6 +52,7 @@
     -webkit-transition: 0.3s;
     transition: 0.3s;
 }
+
 @media (max-width: 30em) {
     .menu__link {
         padding: 1.5rem 0.6rem;
