+++ src/Components/Pages/Cart.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import '../../App.scss';
+const Cart = () => {
+    return (
+        <div class="grid">
+            <div class="grid__item grid__item--sm-span-4">cart</div>
+        </div>
+    );
+};
+
+export default Cart;
